[frontend/허덕형] [6주차] [#321] 마이페이지 위치 추가 페이지와 위치 선택 기능 넣기 (#349)
* [frontend/deokhyeong][#207] 오늘의 날씨 프로젝트 init

* [frontend/deokhyeong][#207] Mobile Web Size Layout Setting

* [frontend/deokhyeong][#207] Login layout

* [frontend/deokhyeong][#207] submitCardForm 컴포넌트와 로그인 페이지 기능 준비

* [frontend/deokhyeong][#207] LogoAndTitle 컴포넌트 작성

* [frontend/deokhyeong][#207] LocalStorage 테스트 아이디 저장

* [frontend/deokhyeong][#207] Fake Axios 만들기

* [frontend/deokhyeong][#207] 홈 페이지 테스트 컴포넌트 작성

* [frontend/deokhyeong][#207] 로그인 기능 추가

* [frontend/deokhyeong][#207] 로그인 실패 메세지 임시 얼럿

* [frontend/deokhyeong][#207] 로그인 실패 메세지 얼럿

* [frontend/deokhyeong][#207] README 업데이트

* [frontend/deokhyeong][#207] 임시적인 token을 이용한 Routing

* [frontend/deokhyeong][#207] Layer 구조 변경으로 인한 Hooks 디렉토리 생성

* [frontend/deokhyeong][#208] 의미 없는 파일,코드 제거

* [frontend/deokhyeong][#208] 회원 가입 페이지 UI 구현

* [frontend/deokhyeong][#208] 회원 가입 중복 이메일 방지

* [frontend/deokhyeong][#208] 회원 가입 페이지 유효성 검사 로직 추가

* [frontend/deokhyeong][#208] 회원 가입 페이지 로컬스토리지 연동 완료

* [frontend/deokhyeong][#208] 구조 변경에 따른 코드 변경 작업

* [frontend/deokhyeong][#207] API 함수 변수 인라인하기

* [frontend/deokhyeong][#207] actions 잘못된 로직 수정

* [frontend/deokhyeong][#207] Error Handle 위치 변경

* [frontend/deokhyeong][#207] TEST코드 검증 세팅

* [frontend/deokhyeong][#208] actions 커밋 컨벤션 반영, issue-207 PR피드백

* [frontend/deokhyeong][#208] logo-title props명 변경

* [frontend/deokhyeong][#208] authStore 모듈명 auth로 변경

* [frontend/deokhyeong][#208] router 발생 시기 변경

* [frontend/deokhyeong][#208] Hooks 폴더 제거

* [frontend/deokhyeong][#208] 불필요한 분리 제거

* [frontend/deokhyeong][#208] 패스워드 확인시 패스워드 Computed를 파라미터로 변경

* [frontend/deokhyeong][#208] 패스워드 확인 Rule 추가, rules vutify 공식문서 컨벤션 따르기

* [frontend/deokhyeong][#208] inputForm 검사 기능 추가, 블러 검사가 기본값

* [frontend/deokhyeong][#207] layout 다시 잡기

* [frontend/deokhyeong][#208] submitFormCard 에러 수정

* [frontend/deokhyeong][#207] vue-cookies 패키지 추가

* [frontend/deokhyeong][#270] 토큰 정보 스토어, 쿠키 저장

* [frontend/deokhyeong][#208] 네비게이션 가드 추가

* [frontend/deokhyeong][#208] 로그인/회원가입 토큰 있을시 리 다이렉트

* [frontend/deokhyeong][#208] 로그인 만료 관련 mock코드 추가

* [frontend/deokhyeong][#208] 의미 없는 파일 제거

* [frontend/deokhyeong][#270] LocalStorage 데이터 반환 객체 변경

* [frontend/deokhyeong][#270] Promise pending 상태인 값 해결

* [frontend/deokhyeong][#270] fake interceptors 작성

* [frontend/deokhyeong][#208] cookies관련 코드 제거

* [frontend/deokhyeong][#208] vue-router 컨벤션 코드로 변경

* [frontend/deokhyeong][#270] Response Interceptor, 비동기 테스트

* [frontend/deokhyeong][#271] 마이페이지 및 네비게이션 세팅

* [frontend/deokhyeong][#270] interceptor 개선 작업

* [frontend/deokhyeong][#271] 마이페이지 위치 설정 UI 작업

* [frontend/deokhyeong][#270] deepClone추가, 인터셉터 관련 코드 수정(fetch관련)

* [frontend/deokhyeong][#270] 인터셉터 파라미터 전달 버그 수정

* [frontend/deokhyeong][#208] 오타 반영

* [frontend/deokhyeong][#208] 불필요한 코드 제거

* [frontend/deokhyeong][#270] 토큰 만료시 재로그인 기준 변경

* [frontend/deokhyeong][#270] try,catch 문때문에 필요없는 Promise reject제거

* [frontend/deokhyeong][#270] 네비게이션 가드 수정

* [frontend/deokhyeong][#270] 스낵바 에러 코드 제거 weather API 로직 추가시 같이 추가 예정

* [frontend/deokhyeong][#270] 라우팅 중복 이동 에러 제거

* [frontend/deokhyeong][#270] input폼 사용시 구글 권장 설정 따르기

* [frontend/deokhyeong][#270] 사용하지 않는 코드 제거

* [frontend/deokhyeong][#270] 테스트용 로그아웃 버튼

* [frontend/deokhyeong][#271] 불필요한 파일 제거

* [frontend/deokhyeong][#270] 테스트용 로그아웃 버튼

* [frontend/deokhyeong][#271] 메인 페이지 네비게이션 바톰 추가

* [frontend/deokhyeong][#271] repair input form 구글 규격에 맞추기

* [frontend/deokhyeong][#271] 버튼 색상 + 기본 메서드 붙일 준비

* [frontend/deokhyeong][#271] input rules 모듈화

* [frontend/deokhyeong][#271] 기본 정보 수정 기능 추가

* [frontend/deokhyeong][#271] 전역 토스트 메세지 컴포넌트 추가

* [frontend/deokhyeong][#270] 흐름에 방해되는 코드들 삭제, jwt encode 불가로 코드 재현 불가한 것도 수정

* [frontend/deokhyeong][#271] 기능성 코드 분리

* [frontend/deokhyeong][#271] 쓸모 없는 코드 제거

* [frontend/deokhyeong][#271] localstorage 코드 분할

* [frontend/deokhyeong][#271] User Store 활용한 코드로 변경

* [frontend/deokhyeong][#271] 유저 스토어에 새로고침 시에 데이터 갱신 건 반영 후 랜더링

* [frontend/deokhyeong][#271] 의미없는 주석 제거

* [frontend/deokhyeong][#321] 셀럭트 리스트 컴포넌트 기초 작업

* [frontend/deokhyeong][#321] 위치 추가 페이지 세팅

* [frontend/deokhyeong][#321] 선택 목록 카드 폼 분리

* [frontend/deokhyeong][#321] 위치 추가 목록 데이터 세팅

* [frontend/deokhyeong][#321] 위치 추가 목록 카드 UI 세팅

* [frontend/deokhyeong][#321] 위치 추가 페이지 타이틀, 백에로우 세팅

* [frontend/deokhyeong][#321] 위치 추가 기능 구현 작업중

* [frontend/deokhyeong][#321] 위치 추가 기능 세팅

* [frontend/deokhyeong][#321] 네비게이션 배경색 수정

* [frontend/deokhyeong][#321] views 페이지들 폴더로 분리(해당 페이지에서만 사용하는 컴포넌트를 위해

* [frontend/deokhyeong][#321] 위치 추가 기능 구현 (선택과 제거 기능 미작업)

* [frontend/deokhyeong][#321] 위치 추가 기능 (제거, 선택)기능 추가

* [frontend/deokhyeong][#270] 네비게이션 가드에서 사이드 이펙트 가능성 없애기

* [frontend/deokhyeong][#270] 네비게이션 가드 불명확한 함수명, 변수명 변경

* [frontend/deokhyeong][#270] Vuetify Text field  컨벤션 커스텀 되돌리기

* [frontend/deokhyeong][#270] axios mocking 기본 세팅 구현

* [frontend/deokhyeong][#270] 기간 만료 검사 utils추가

* [frontend/deokhyeong][#270] 명령 함수와 질의  함수 분리하기

* [frontend/deokhyeong][#271] 유저 스토어에 새로고침 시에 데이터 갱신 건 반영 후 렌더링

* [frontend/deokhyeong][#271] 불필요한 파일 제거

* [frontend/deokhyeong][#270] 명령 함수와 질의  함수 분리하기

* [frontend/deokhyeong][#270] shared => utils로 변경

* [frontend/deokhyeong][#270] shared 디렉토리 제거

* [frontend/deokhyeong][#270] 사용하지 않는 error 디렉토리 제거

* [frontend/deokhyeong][#270] fakeHttp lib 디렉토리로 분리

* [frontend/deokhyeong][#270] view파일 index.vue로 변경 작업

* [frontend/deokhyeong][#271] 네비게이션 Bottom UI 수정

* [frontend/deokhyeong][#270] view파일 index.vue로 변경 작업

* [frontend/deokhyeong][#271] shallow Equal 빠진거 추가

* [frontend/deokhyeong][#321] 불필요한 경로 탐색 제거

* [frontend/deokhyeong][#271] 에러 처리 추가

* [frontend/deokhyeong][#271] isEqual => isShallowEqual로 교체

* [frontend/deokhyeong][#271] 네비게이션 복구

* [frontend/deokhyeong][#271] 에러 로그, 개발환경에서는 볼 수 있도록 처리

* [frontend/deokhyeong][#321] 병렬 에러 로그, 개발환경에서 볼 수 있도록 처리

* [frontend/deokhyeong][#321] 타입 에러 수정

* [frontend/deokhyeong][#321] 오타 수정

* [frontend/deokhyeong][#271] catchAsync 제거, 에러처리 작업 방향 변경

* [frontend/deokhyeong][#271] App 컴포넌트 알아보기 힘든 변수명 및 코드 흐름 수정

* [frontend/deokhyeong][#271] fetchRepairUserInfo => patchRepairUserInfo로 너무 과한 이름 수정

* [frontend/deokhyeong][#271] 에러 메세지 파싱 준비

* [frontend/deokhyeong][#271] 토스트 메세지 동작 방식 변경, error store의 도입

* [frontend/deokhyeong][#271] 반복되고 읽기 힘든 계산식 함수화

* [frontend/deokhyeong][#271] state 자체를 참조로 사용한 코드 제거

* [frontend/deokhyeong][#271] error 반환 기본메세지 수정

* [frontend/deokhyeong][#271] 비밀번호만 수정 못하게하는 버그 수정

* [frontend/deokhyeong][#321] 신규 유저 생성시 유저 정보 포멧  수정

* [frontend/deokhyeong][#271] 남아있는 토스트 open형식 제거

* [frontend/deokhyeong][#271] 회원가입 비동기 처리 안 된 코드 반영

* [frontend/deokhyeong][#321] catch Async 걷어내기 적용 진행

* [frontend/deokhyeong][#321] catch Async 걷어내기 완료

* [frontend/deokhyeong][#271] 기본 오류 메세지 세팅

* [frontend/deokhyeong][#271] 로그아웃 시 동작 변경

* [frontend/deokhyeong][#271] 토스트 메세지 컴포넌트 세팅

* [frontend/deokhyeong][#321] 린트 에러 수정

* [frontend/deokhyeong][#271] localstorage user 초기화가 무조건 발생하는 버그 수정

* [frontend/deokhyeong][#321] 불필요한 토스트메세지 컴포넌트 제거

* [frontend/deokhyeong][#321] 위치 추가 기능 권한 설정 및 초기화 기능 추가

* [frontend/deokhyeong][#321] 성공 메세지에 관한 메서드 추가 및 세팅

* [frontend/deokhyeong][#321] 마이 페이지 안내 문구 변경

* [frontend/deokhyeong][#321] locationAdd > created >  async await 코드 추가

* [frontend/deokhyeong][#321] 부연설명이 필요한 코드에 주석 추가

* [frontend/deokhyeong][#321] 변수명 변경 작업 및 uuid를 key id로 사용하기

* [frontend/deokhyeong][#321] 통신 mocking으로 보이기 위한 axios 인스턴스화 작업

* [frontend/deokhyeong][#321] error 스토어의 역할보단 alert의 역할에 가까워 alert store 따로 추가

* [frontend/deokhyeong][#321] alert Store 코드 연결

* [frontend/deokhyeong][#321] 불필요한 코드 제거

* [frontend/deokhyeong][#321] Utils에 남은 catchAsync제거

+++ frontend/deokhyeong/weather/mocks/fakeServer/fetchAddLocation.js
@@ -1,8 +1,16 @@
-import utils from '@/utils'
-
-const patchRepairUserInfo = (headers, body) => {
+const fetchAddLocation = (headers, body) => {
   const users = JSON.parse(localStorage.getItem('users'))
   const targetIndex = users.findIndex((_user) => _user.accessToken === headers?.Authentication)
+
+  if (!body.location) {
+    const errorResponse = Promise.reject({
+      status: 400,
+      data: {
+        message: 'BadRequest',
+      },
+    })
+    return errorResponse
+  }
 
   if (!headers?.Authentication) {
     const errorResponse = Promise.reject({
@@ -25,7 +33,8 @@ const patchRepairUserInfo = (headers, body) => {
   }
 
   const targetUser = users[targetIndex]
-  if (utils.isExpiredTime(targetUser.expire)) {
+  const isExpiredDate = (targetUser.expire - new Date().getTime()) < 0
+  if (isExpiredDate) {
     const expiredResponse = Promise.reject({
       status: 401,
       data: {
@@ -35,36 +44,32 @@ const patchRepairUserInfo = (headers, body) => {
     return expiredResponse
   }
 
-  const prevEmail = users[targetIndex].email
-  const isSameEmailWithPrevEmail = prevEmail === body.email
-  if (!isSameEmailWithPrevEmail) {
-    const isEmailOverLap = users.find((_user) => _user.email === body.email)
-    if (isEmailOverLap) {
-      const emailOverLapResponse = Promise.reject({
-        status: 400,
-        data: {
-          message: '이메일 중복',
-        },
-      })
-
-      return emailOverLapResponse
-    }
+  const isExistedLocation = targetUser.bookmarkLocations.find(
+    (bookmark) => bookmark.location === body.location,
+  )
+  if (isExistedLocation) {
+    const existedLocationResponse = Promise.reject({
+      status: 400,
+      data: {
+        message: '이미 북마크에 존재하는 위치입니다.',
+      },
+    })
+    return existedLocationResponse
   }
 
-  targetUser.email = body.email
-  targetUser.password = body.newPassword
+  targetUser.bookmarkLocations.push(body)
+  targetUser.selectedLocation = body
   users[targetIndex] = targetUser
   localStorage.setItem('users', JSON.stringify(users))
   const successResponse = Promise.resolve({
     status: 200,
     data: {
       message: 'Success',
-      ...targetUser,
-      repairTag: Math.random(), // repairApi로 수정한 계정만 가짐
+      bookmarkLocations: targetUser.bookmarkLocations,
+      selectedLocation: targetUser.selectedLocation,
     },
   })
-
   return successResponse
 }
 
-export default patchRepairUserInfo
+export default fetchAddLocation
